requirements.txt
streamlit
openai
deepgram-sdk
pandas
openpyxl
gspread
google-auth
anthropic>=0.25.0
